more formatting

+++ array/doubleStack.c
@@ -120,7 +120,7 @@ void display1(doubleStack d) {
     printf("Empty stack\n");
     return;
   }
-  for (short i = d.top1; i > -1; i--)
+  for (int i = d.top1; i > -1; i--)
     printf("%d\n", d.arr[i]);
 }
 
@@ -129,6 +129,6 @@ void display2(doubleStack d) {
     printf("Empty stack\n");
     return;
   }
-  for (short i = d.top2; i < MAX; i++)
+  for (int i = d.top2; i < MAX; i++)
     printf("%d\n", d.arr[i]);
 }
